pyproject.toml
[project]
name = "poke-scrape"
version = "0.1.0"
description = ""
authors = [
    {name = "Michal Wodnicki",email = "michalwodnicki23@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.129.0,<0.130.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "requests (>=2.32.5,<3.0.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "flask (>=3.1.2,<4.0.0)",
    "gunicorn (>=25.1.0,<26.0.0)",
    "pokemontcgsdk (>=3.4.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
[dependency-groups]
dev = [
    "ipykernel (>=7.2.0,<8.0.0)"
]
